feat: generate horizontal stacked percent bar charts with chartjs
Prompt: The first column of the csv file contains value for grouping. The second and third columns contain answers to questions. Possible answers are 1, 2, 3 and 4. The global idea is to be able to see the proportion for each answers depending on the group. Then, I want the app to generate a graph for each questions (second and third columns of the csv file). The generated graph must be displayed on the page. The graph must be generated with chartjs library. The graph must be an horizontal stacked bar. The answers 1, 2, 3 and 4 must generate a stacked bar taking the full width of the graph, with different colors for each value and a size proportional to the number of answers. I want one stacked bars for each group (first column of the csv file) and one stacked bar for all the answers without group filtering.