refactor: split component in 3 dedicated component
Handle xls and xlsx in addition of csv files.
prompt: Is it possible to split src/components/CSVUpload.js into several component. I would like to split the extraction logic from the chart generation logic and from the data processing. Furthermore, I would like to be able to work with xls and xlsx files in addition of csv files.