feat: integrate result from previous surveys
Prompt: Great! Now instead of uploading only one file corresponding to a survey, I want to be able to upload 2 other files corresponding to the previous occurences of the survey. Then, I want a chart for each questions and for each group and overall. A chart for a group must contain one stacked bar for the total of the current survey,  one stacked bar for the current survey for the current group, and one stacked bar for each previous surveys for the current group. A chart for overall result must contain one stacked bar for the total of the current survey and one stack bar for each total of the previous surveys. As label, previous survey must be call "S -1" and "S -2".

Prompt for correction: The charts for each group are correct. They contain the overall, the current, the S -1 and the S -2 bars. But the overall charts are not correct. They contain an overall bar which must not be present. And the current, S -1 and S -2 bar are empty. For the overall charts, I want
- a current stacked bar containing result for the current chart without grouping
- a S -1 stacked bar containing result for the S -1 chart without grouping
- a S -2 stacked bar containing result for the S -2 chart without grouping

Prompt for modification: The overall charts must also contain the current result for each group.
- current
- S -1
- S -2
- Group A
- Group B
- etc...