feat: handle several and configurable grouping column and pdf generation
Prompt: I want now to handle several improvments.

I want the app to handle several columns for grouping. When uploading the current survey, the app must display the header and the first 2 lines of the uploaded files. An horizontal scrollbar must be used to easily view the whole header. Then the user must be able to click on the header to determine which columns are used for grouping. The selected columns must be used for grouping. All other columns are by default considered as question columns for which the app will generate charts.

The app must generate overall charts and then generate charts for each group of each grouping for all questions. I want the app to display first all the overall charts order by questions ordering. Then all charts for the first grouping column ordered by group and then by questions ordering. Then all charts for the second grouping, etc...

I want the app to display a download button. A click on the button must lead to the download of a zip file. The zip file must contain one pdf file called overall.pdf containing all the overall charts, and one directory per grouping column. Each directory per grouping column must contain one pdf file for each group called with the name of the group and containing all the charts of this group.

I want the presence of previous surveys to be optional. The user must be able to download from 0 to 3 previous surveys. A generate button must be present to trigger the validation of the grouping columns definitino, the data extraction and the charts generation.

Prompt for debugging: I've got the following error: src/App.js
  Line 38:22:   'processSurveyData' is not defined  no-undef
  Line 39:102:  'processSurveyData' is not defined  no-undef

Prompt for debugging: Well! I now get the following error when I upload the current survey: onFileParsed is not a function
TypeError: onFileParsed is not a function
    at Object.complete (http://localhost:3000/static/js/bundle.js:580:15)
    at p.parseChunk (http://localhost:3000/static/js/bundle.js:31400:339)
    at p._nextChunk (http://localhost:3000/static/js/bundle.js:31481:109)
    at p.stream (http://localhost:3000/static/js/bundle.js:31476:26)
    at Object.parse (http://localhost:3000/static/js/bundle.js:31249:18)
    at reader.onload (http://localhost:3000/static/js/bundle.js:574:60)

Prompt for product correction but the answer was not correct and I got more and more bug trying to continue: I want 3 products corrections:

I want the questions to be present in the generated PDF file just on the top of the chart.

I finally want overall charts for each grouping columns. The overall charts must only contains the group of the current grouping column and not the other one.

I want group to only depend on one grouping column. Currently the app generate groups that are intersection of several grouping columns.